+++ js/addAndEditArticleAjax.js
@@ -1,0 +1,5 @@
+/**
+ *
+ *  todo: Сделать добавление статьи через ajax.
+ *
+ */
