+++ js/deleteArticleAjax.js
