+++ src/service/Log.php
@@ -8,19 +8,11 @@
  * @link        https://github.com/pers1307/Blog_v_2.0
  */
 
-namespace pers1307\blog\services;
+namespace pers1307\blog\service;
 
 use Monolog\Logger;
 use Monolog\Handler\StreamHandler;
 
-/**
- *
- * Мне логи к каким событиям привязывать?
- * Ко всем эксепшенам? Или там ещё и к системным какм нибудь нейтральным?
- *
- * Class Log
- * @package pers1307\blog\services
- */
 class Log
 {
     /** @var Log */
@@ -32,7 +24,7 @@ class Log
     private function __construct()
     {
         $this->log = new Logger('log');
-        $today = date('Y-m-d');
+        $today = date('d-m-Y');
         $path = 'app/logs/' . $today . '.log';
         $this->log->pushHandler(new StreamHandler($path, Logger::ERROR));
     }
