+++ views/articleDesk.html
@@ -8,51 +8,8 @@
 
 <div id="ContentControlPanel" class="container">
     <div class="row">
-        <div class="col-md-12">
-            <div id="New-post">
-                <h3>Добавить новую статью в блог!</h3>
-                <form role="form" method="post" action="" enctype="multipart/form-data">
-                    {% if errorAddArticle['CodeError'] != 0 %}
-                        <div class="form-group alert-danger col-md-12">
-                            <h4>{{ errorAddArticle['TextError'] }}</h4>
-                        </div>
-                    {% endif %}
-
-                    <div class="form-group col-md-8 {{ (errorAddArticle['CodeError'] == '1') ? 'has-error' }}">
-                        <label for="exampleInputEmail1">Название статьи</label>
-                        <input type="text" class="form-control" placeholder="Название статьи" name="newArticleName" value="{{ (errorAddArticle['CodeError'] != '0') ? errorAddArticle['Article'].getName() }}">
-                    </div>
-                    <div class="form-group col-md-4 {{ (errorAddArticle['CodeError'] == '2') ? 'has-error' }}">
-                        <label for="exampleInputPassword1">Автор статьи</label>
-                        <input type="text" class="form-control" placeholder="Автор статьи" value="{{ (errorAddArticle['CodeError'] != '0') ? errorAddArticle['Article'].getAuthor() : 'Перескоков Юрий' }}" name="newArticleAuthor">
-                    </div>
-
-                    <div class="form-group col-md-12" {{ (errorAddArticle['CodeError'] == '3') ? "style='outline: 1px solid #A90006'" }} >
-
-                        {% if errorAddArticle['CodeError'] != 0 %}
-                            <label for="exampleInputFile">Текущая картинка: {{ errorAddArticle['Article'].getPathImage() }} </label>
-                            <p><label for="exampleInputFile">Добавьте её снова!</label></p>
-                        {% endif %}
-
-                        <label for="exampleInputFile">Хочу добавить картинку!</label>
-                        <input type="file" name="newArticleImage[]" multiple="multiple">
-                    </div>
-
-                    <div class="form-group col-md-12 {{ (errorAddArticle['CodeError'] == '4') ? 'has-error' }}">
-                        <label for="exampleInputFile">Текст статьи: </label>
-                        <textarea id="newArticle" class="form-control" rows="6" name="newArticleText" style="resize: none">{{ (errorAddArticle['CodeError'] != '0') ? errorAddArticle['Article'].getText() }}</textarea>
-                    </div>
-
-                    <div class="form-group">
-                        <button id="ForSub" type="submit" class="btn btn-primary" name="sub"><i class="fa fa-check"></i>Отправить</button>
-                    </div>
-                </form>
-            </div>
-        </div>
-
-
         <div id='WrapAccordion'>
-            <h3>Редактирование существующих статей</h3>
+            <h3>Редактирование статей</h3>
             <div class="panel-group" id="accordion">
                 {% for article in articles %}
                     <div class="panel panel-default" id="idPost{{ article.getId() }}">
@@ -62,7 +19,7 @@
                                     Дата {{ article.getCreatedAt() }} Название статьи: {{ article.getName() }}
                                 </a>
                                 <a href='?Delete={{ article.getId() }}' class='btn btn-danger btn-xs delete' data-delete='{{ article.getId() }}'><i class="fa fa-times fa-2x"></i></a>
-                                <a href='/EditArticle?Edit={{ article.getId() }}' class='btn btn-info btn-xs'><i class="fa fa-pencil-square-o fa-2x"></i></a>
+                                <a href='/article/{{ article.getId() }}' class='btn btn-info btn-xs'><i class="fa fa-pencil-square-o fa-2x"></i></a>
                             </h4>
                         </div>
                         <div id="{{ article.getId() }}" class="panel-collapse collapse">
@@ -76,9 +33,10 @@
                 {% else %}
                     <h1>В данный момент в блоге нет статей!</h1>
                 {% endfor %}
+                <br />
+                <a href='/article/0' class='btn btn-primary btn-xs addArticle'><i class="fa fa-plus fa-2x"></i> Добавить статью</a>
+                <br />
             </div>
-            <a href='/article' class='btn btn-primary btn-xs'><i class="fa fa-plus fa-2x"></i> Добавить статью</a>
-            <br />
         </div>
     </div>
 </div>
